2006-07-10  Maurits  <m.rijk@chello.nl>
	* lib/Procedure.cs: convert bool to Int32 if needed in RunProcedure

	* plug-ins/PhotoshopActions/AddMaskEvent.cs: implemented

	* plug-ins/PhotoshopActions/SelectAllEvent.cs
	* plug-ins/PhotoshopActions/SelectNoneEvent.cs
	* plug-ins/PhotoshopActions/SelectRectangleEvent.cs
	* plug-ins/PhotoshopActions/SelectEllipseEvent.cs: new files

	* plug-ins/PhotoshopActions/Makefile.am: changed accordingly

	* lib/Mask.cs: derive Mask from Drawable

	* lib/EllipseSelectTool.cs: new files
	* lib/Makefile.am: changed accordingly

+++ lib/Selection.cs
@@ -131,7 +131,7 @@ namespace Gimp
         }
     }
 
-    public void Feather(int radius)
+    public void Feather(double radius)
     {
       if (!gimp_selection_feather (_imageID, radius))
         {
